+++ model/atmosphere/advection/tests/advection_stencil_tests/test_mo_advection_traj_btraj_compute_o1_dsl.py
@@ -69,7 +69,7 @@ class TestMoAdvectionTrajBtrajComputeO1Dsl(StencilTest):
         dual_normal_cell_2 = np.expand_dims(dual_normal_cell_2, axis=-1)
 
         p_cell_idx = np.where(lvn_pos, cell_idx[:, 0], cell_idx[:, 1])
-        p_cell_rel_idx_dsl = np.where(lvn_pos, int32(0), int32(1))
+        p_cell_rel_idx_dsl = np.where(lvn_pos, 0, 1)
         p_cell_blk = np.where(lvn_pos, cell_blk[:, 0], cell_blk[:, 1])
 
         z_ntdistv_bary_1 = -(
@@ -147,8 +147,8 @@ class TestMoAdvectionTrajBtrajComputeO1Dsl(StencilTest):
             p_distv_bary_1=p_distv_bary_1,
             p_distv_bary_2=p_distv_bary_2,
             p_dthalf=p_dthalf,
-            horizontal_start=int32(0),
+            horizontal_start=0,
             horizontal_end=int32(grid.num_edges),
-            vertical_start=int32(0),
+            vertical_start=0,
             vertical_end=int32(grid.num_levels),
         )
