+++ model/atmosphere/advection/tests/advection_stencil_tests/test_vert_adv_stencil_01.py
@@ -40,7 +40,7 @@ class TestVertAdvStencil01(StencilTest):
         p_dtime: float,
         **kwargs,
     ) -> np.array:
-        if ivadv_tracer != int32(0):
+        if ivadv_tracer != 0:
             tracer_new = np.where(
                 (iadv_slev_jt <= k),
                 (
@@ -69,8 +69,8 @@ class TestVertAdvStencil01(StencilTest):
         rhodz_new = random_field(grid, CellDim, KDim)
         k = as_field((KDim,), np.arange(grid.num_levels, dtype=int32))
         p_dtime = np.float64(5.0)
-        ivadv_tracer = int32(1)
-        iadv_slev_jt = int32(4)
+        ivadv_tracer = 1
+        iadv_slev_jt = 4
         tracer_new = zero_field(grid, CellDim, KDim)
         return dict(
             tracer_now=tracer_now,
