+++ model/atmosphere/diffusion/tests/diffusion_stencil_tests/test_truly_horizontal_diffusion_nabla_of_theta_over_steep_points.py
@@ -139,8 +139,8 @@ class TestTrulyHorizontalDiffusionNablaOfThetaOverSteepPoints(StencilTest):
             theta_v=theta_v,
             z_temp=z_temp,
             vcoef=vcoef_new,
-            horizontal_start=int32(0),
+            horizontal_start=0,
             horizontal_end=int32(grid.num_cells),
-            vertical_start=int32(0),
+            vertical_start=0,
             vertical_end=int32(grid.num_levels),
         )
