+++ model/atmosphere/dycore/tests/dycore_stencil_tests/test_copy_cell_kdim_field_to_vp.py
@@ -42,8 +42,8 @@ class TestCopyCellKdimFieldToVp(StencilTest):
         return dict(
             field=field,
             field_copy=field_copy,
-            horizontal_start=int32(0),
+            horizontal_start=0,
             horizontal_end=int32(grid.num_cells),
-            vertical_start=int32(0),
+            vertical_start=0,
             vertical_end=int32(grid.num_levels),
         )
