+++ model/atmosphere/dycore/tests/dycore_stencil_tests/test_update_mass_volume_flux.py
@@ -64,8 +64,8 @@ class TestUpdateMassVolumeFlux(StencilTest):
             mass_flx_ic=mass_flx_ic,
             vol_flx_ic=vol_flx_ic,
             r_nsubsteps=r_nsubsteps,
-            horizontal_start=int32(0),
+            horizontal_start=0,
             horizontal_end=int32(grid.num_cells),
-            vertical_start=int32(0),
+            vertical_start=0,
             vertical_end=int32(grid.num_levels),
         )
